Split CLI config handling out of ConfigBuilder
Extract the CLI-specific configuration logic (option/command handling and
config-file resolution: home, base dir, local, and user config files) from
`ConfigBuilder` into a new `ConfigCmdAdapter` that wraps it. `ConfigBuilder`
is left as a lean, `@CompileStatic` runtime builder that consumes an explicit
list of config entries, and all call sites are migrated to `ConfigCmdAdapter`
(CmdRun, CmdConfig, CmdClean, CmdNode, CmdFs, CmdLineage, CmdAuth, the module
commands, PluginAbstractExec, K8sDriverLauncher, and the nf-console/nf-tower
plugins). `DefaultRemoteModuleResolver` is updated for the new build() API.

This prepares the runtime config builder to be decoupled from the CLI, and
keeps the forthcoming CLI/runtime module split a mechanical move.

Two intentional behavior changes come with the rewrite (both covered by new
tests in ConfigBuilderTest):

- `session` is no longer created as a guaranteed top-level config scope. It
  is not a config scope, so code must not rely on it existing when unset.
- The v2 config parser no longer exposes environment variables to the config
  binding (they remain available to the v1 parser). The v2 parser does not
  support implicit environment variables.
