+++ modules/nextflow/src/main/groovy/nextflow/cli/CmdConfig.groovy
@@ -26,6 +26,7 @@ import groovy.transform.PackageScope
 import groovy.util.logging.Slf4j
 import nextflow.NF
 import nextflow.config.ConfigBuilder
+import nextflow.config.ConfigCmdAdapter
 import nextflow.config.ConfigValidator
 import nextflow.exception.AbortOperationException
 import nextflow.plugin.Plugins
@@ -113,12 +114,13 @@ class CmdConfig extends CmdBase {
         final builder = new ConfigBuilder()
                 .setShowClosures(true)
                 .setStripSecrets(true)
-                .showMissingVariables(true)
+                .setShowMissingVariables(true)
+
+        final config = new ConfigCmdAdapter(builder)
                 .setOptions(launcher.options)
                 .setBaseDir(base)
                 .setCmdConfig(this)
-
-        final config = builder.buildConfigObject()
+                .buildConfigObject()
 
         // -- validate config options
         if( NF.isSyntaxParserV2() ) {
