+++ modules/nextflow/src/main/groovy/nextflow/module/DefaultRemoteModuleResolver.groovy
@@ -18,6 +18,7 @@ package nextflow.module
 
 import groovy.transform.CompileStatic
 import groovy.util.logging.Slf4j
+import nextflow.Const
 import nextflow.Global
 import nextflow.config.ConfigBuilder
 
@@ -47,7 +48,7 @@ class DefaultRemoteModuleResolver implements RemoteModuleResolver {
     @Override
     Path resolve(String moduleName, Path projectDir) {
         final baseDir = projectDir ?: Path.of('.').toAbsolutePath()
-        final config = Global.config ?: new ConfigBuilder().setBaseDir(baseDir).build()
+        final config = Global.config ?: buildConfig(baseDir)
         final registryConfig = config.navigate('registry') as RegistryConfig
 
         // Create module resolver
@@ -72,5 +73,20 @@ class DefaultRemoteModuleResolver implements RemoteModuleResolver {
     @Override
     int getPriority() {
         return 0  // Default implementation has lowest priority
+    }
+
+    /**
+     * Build a minimal config from the default config files (Nextflow home and the
+     * project base directory) when no global config is available.
+     */
+    protected static Map buildConfig(Path baseDir) {
+        final files = new ArrayList<Path>()
+        final home = Const.APP_HOME_DIR.resolve('config')
+        if( home.exists() )
+            files.add(home)
+        final base = baseDir.resolve('nextflow.config')
+        if( base.exists() )
+            files.add(base)
+        return new ConfigBuilder().setBaseDir(baseDir).build(files)
     }
 }
