+++ modules/nextflow/src/test/groovy/nextflow/config/ConfigBuilderTest.groovy
@@ -20,11 +20,6 @@ import java.nio.file.Files
 import java.nio.file.Path
 
 import nextflow.SysEnv
-import nextflow.cli.CliOptions
-import nextflow.cli.CmdConfig
-import nextflow.cli.CmdNode
-import nextflow.cli.CmdRun
-import nextflow.cli.Launcher
 import nextflow.exception.AbortOperationException
 import nextflow.exception.ConfigParseException
 import nextflow.extension.FilesEx
@@ -39,32 +34,6 @@ import spock.lang.Unroll
  */
 class ConfigBuilderTest extends Specification {
 
-    def setupSpec() {
-        SysEnv.push([:])
-    }
-
-    def cleanupSpec() {
-        SysEnv.pop()
-    }
-
-    ConfigObject configWithParams(Path file, Map runOpts, Path baseDir=null) {
-        def run = new CmdRun(runOpts)
-        return new ConfigBuilder()
-            .setOptions(new CliOptions())
-            .setCmdRun(run)
-            .setCliParams(run.parsedParams(ConfigBuilder.getConfigVars(baseDir, null)))
-            .buildGivenFiles(file)
-    }
-
-    ConfigObject configWithParams(Map config, Map runOpts, Map cliOpts=[:]) {
-        def run = new CmdRun(runOpts)
-        return new ConfigBuilder(config)
-            .setOptions(new CliOptions(cliOpts))
-            .setCmdRun(run)
-            .setCliParams(run.parsedParams(ConfigBuilder.getConfigVars(null, null)))
-            .build()
-    }
-
     def 'build config object' () {
 
         setup:
@@ -72,7 +41,7 @@ class ConfigBuilderTest extends Specification {
         def builder = [:] as ConfigBuilder
 
         when:
-        def config = builder.buildConfig0(env,null)
+        def config = builder.build(env,null)
 
         then:
         ('PATH' in config.env )
@@ -120,8 +89,8 @@ class ConfigBuilderTest extends Specification {
 
         when:
         def env = SysEnv.get()
-        def config1 = builder.buildConfig0(env, [text1])
-        def config2 = builder.buildConfig0(env, [text1, text2])
+        def config1 = builder.build(env, [text1])
+        def config2 = builder.build(env, [text1, text2])
 
         // note: configuration object can be modified like any map
         config2.env ['ZZZ'] = '99'
@@ -173,10 +142,9 @@ class ConfigBuilderTest extends Specification {
         params.test = 2
         '''
 
-
         when:
         def env = SysEnv.get()
-        def config1 = builder.buildConfig0(env, [text1])
+        def config1 = builder.build(env, [text1])
 
         then:
         config1.task.field1 == 1
@@ -208,1303 +176,13 @@ class ConfigBuilderTest extends Specification {
         '''
 
         when:
-        def cfg = builder.buildConfig0([:], [text])
+        def cfg = builder.build([:], [text])
         then:
         cfg.params.p == '/base/path/1'
         cfg.params.q == '/base/path/2'
         cfg.params.x == '/base/path/3'
         cfg.params.y == "${Path.of('.').toRealPath()}/4"
 
-    }
-
-    def 'CLI params should override the ones defined in the config file' () {
-        setup:
-        def file = Files.createTempFile('test',null)
-        file.text = '''
-        params {
-          alpha = 'x'
-        }
-        params.beta = 'y'
-        params.delta = 'Foo'
-        params.gamma = params.alpha
-        params {
-            omega = 'Bar'
-        }
-
-        process {
-          publishDir = [path: params.alpha]
-        }
-        '''
-        when:
-        def result = configWithParams(file, [params: [alpha: 'Hello', beta: 'World', omega: 'Last']])
-
-        then:
-        result.params.alpha == 'Hello'  // <-- params defined as CLI options override the ones in the config file
-        result.params.beta == 'World'   // <--   as above
-        result.params.gamma == 'Hello'  // <--   as above
-        result.params.omega == 'Last'
-        result.params.delta == 'Foo'
-        result.process.publishDir == [path: 'Hello']
-
-        cleanup:
-        file?.delete()
-    }
-
-    def 'CLI params should override the ones defined in the config file [2]' () {
-        setup:
-        def file = Files.createTempFile('test',null)
-        file.text = '''
-        params {
-          alpha = 'x'
-          beta = 'y'
-          delta = 'Foo'
-          gamma = params.alpha
-          omega = 'Bar'
-        }
-
-        process {
-          publishDir = [path: params.alpha]
-        }
-        '''
-        when:
-        def result = configWithParams(file, [params: [alpha: 'Hello', beta: 'World', omega: 'Last']])
-
-        then:
-        result.params.alpha == 'Hello'  // <-- params defined as CLI options override the ones in the config file
-        result.params.beta == 'World'   // <--   as above
-        result.params.gamma == 'Hello'  // <--   as above
-        result.params.omega == 'Last'
-        result.params.delta == 'Foo'
-        result.process.publishDir == [path: 'Hello']
-
-        cleanup:
-        file?.delete()
-    }
-
-
-    def 'CLI params should override the ones in one or more config files' () {
-        given:
-        def folder = File.createTempDir()
-        def configMain = new File(folder,'nextflow.config').absoluteFile
-        def snippet1 = new File(folder,'config1.txt').absoluteFile
-        def snippet2 = new File(folder,'config2.txt').absoluteFile
-
-
-        configMain.text = """
-        process.name = 'alpha'
-        params.one = 'a'
-        params.xxx = 'x'
-        includeConfig "$snippet1"
-        """
-
-        snippet1.text = """
-        params.two = 'b'
-        params.yyy = 'y'
-
-        process.cpus = 4
-        process.memory = '8GB'
-
-        includeConfig("$snippet2")
-        """
-
-        snippet2.text = '''
-        params.three = 'c'
-        params.zzz = 'z'
-
-        process { disk = '1TB' }
-        process.resources.foo = 1
-        process.resources.bar = 2
-        '''
-
-        when:
-        def config = configWithParams(configMain.toPath(), [params: [one: '1', two: 'dos', three: 'tres']])
-
-        then:
-        config.params.one == '1'
-        config.params.two == 'dos'
-        config.params.three == 'tres'
-        config.process.name == 'alpha'
-        config.params.xxx == 'x'
-        config.params.yyy == 'y'
-        config.params.zzz == 'z'
-
-        config.process.cpus == 4
-        config.process.memory == '8GB'
-        config.process.disk == '1TB'
-        config.process.resources.foo == 1
-        config.process.resources.bar == 2
-
-        cleanup:
-        folder?.deleteDir()
-    }
-
-    def 'should include config with params' () {
-        given:
-        def folder = File.createTempDir()
-        def configMain = new File(folder,'nextflow.config').absoluteFile
-        def snippet1 = new File(folder,'igenomes.config').absoluteFile
-
-
-        configMain.text = '''
-        includeConfig 'igenomes.config'
-        '''
-
-        snippet1.text = '''
-        params {
-          genomes {
-            'GRCh37' {
-              fasta = "${params.igenomes_base}/genome.fa"
-              bwa   = "${params.igenomes_base}/BWAIndex/genome.fa"
-            }
-          }
-        }
-        '''
-
-        when:
-        def config = configWithParams(configMain.toPath(), [params: [igenomes_base: 'test']])
-
-        then:
-        config.params.genomes.GRCh37 == [fasta:'test/genome.fa', bwa:'test/BWAIndex/genome.fa']
-
-        cleanup:
-        folder?.deleteDir()
-    }
-
-
-    def 'should fetch the config path from env var' () {
-        given:
-        def folder = File.createTempDir()
-        def configMain = new File(folder,'my.config').absoluteFile
-
-        configMain.text = """
-        process.name = 'alpha'
-        params.one = 'a'
-        params.two = 'b'
-        """
-
-        // relative path to current dir
-        when:
-        def config = new ConfigBuilder(env: [NXF_CONFIG_FILE: 'my.config']) .setCurrentDir(folder.toPath()) .build()
-        then:
-        config.params.one == 'a'
-        config.params.two == 'b'
-        config.process.name == 'alpha'
-
-        // absolute path
-        when:
-        config = new ConfigBuilder(env: [NXF_CONFIG_FILE: configMain.toString()]) .build()
-        then:
-        config.params.one == 'a'
-        config.params.two == 'b'
-        config.process.name == 'alpha'
-
-        // default should not find it
-        when:
-        config = new ConfigBuilder() .build()
-        then:
-        config.params == [:]
-
-        cleanup:
-        folder?.deleteDir()
-    }
-
-    def 'CLI params should override params in profiles' () {
-
-        setup:
-        def file = Files.createTempFile('test',null)
-        file.text = '''
-        params.alpha = 'a'
-        params.beta = 'b'
-        params.delta = 'Foo'
-        params.gamma = params.alpha
-
-        params {
-            genomes {
-                'GRCh37' {
-                    bed12   = '/data/genes.bed'
-                    bismark = '/data/BismarkIndex'
-                    bowtie  = '/data/genome'
-                }
-            }
-        }
-
-        profiles {
-            first {
-                params.alpha = 'Alpha'
-                params.omega = 'Omega'
-                params.gamma = 'First'
-                process.name = 'Bar'
-            }
-
-            second {
-                params.alpha = 'xxx'
-                params.gamma = 'Second'
-                process {
-                    publishDir = [path: params.alpha]
-                }
-            }
-        }
-        '''
-
-        when:
-        def config = configWithParams(file, [params: [alpha: 'AAA', beta: 'BBB']])
-        then:
-        config.params.alpha == 'AAA'
-        config.params.beta == 'BBB'
-        config.params.delta == 'Foo'
-        config.params.gamma == 'AAA'
-        config.params.genomes.'GRCh37'.bed12 == '/data/genes.bed'
-        config.params.genomes.'GRCh37'.bismark == '/data/BismarkIndex'
-        config.params.genomes.'GRCh37'.bowtie  == '/data/genome'
-
-        when:
-        config = configWithParams(file, [params: [alpha: 'AAA', beta: 'BBB'], profile: 'first'])
-        then:
-        config.params.alpha == 'AAA'
-        config.params.beta == 'BBB'
-        config.params.delta == 'Foo'
-        config.params.gamma == 'First'
-        config.process.name == 'Bar'
-        config.params.genomes.'GRCh37'.bed12 == '/data/genes.bed'
-        config.params.genomes.'GRCh37'.bismark == '/data/BismarkIndex'
-        config.params.genomes.'GRCh37'.bowtie  == '/data/genome'
-
-        cleanup:
-        file?.delete()
-    }
-
-    def 'params-file should override params in the config file' () {
-        setup:
-        def baseDir = Path.of('/my/base/dir')
-        and:
-        def paramsFile = Files.createTempFile('test', '.yml')
-        paramsFile.text = '''
-            alpha: "Hello"
-            beta: "World"
-            omega: "Last"
-            theta: "${baseDir}/something"
-            '''.stripIndent()
-        and:
-        def file = Files.createTempFile('test',null)
-        file.text = '''
-        params {
-          alpha = 'x'
-        }
-        params.beta = 'y'
-        params.delta = 'Foo'
-        params.gamma = params.alpha
-        params {
-            omega = 'Bar'
-        }
-
-        process {
-          publishDir = [path: params.alpha]
-        }
-        '''
-        when:
-        def result = configWithParams(file, [paramsFile: paramsFile], baseDir)
-
-        then:
-        result.params.alpha == 'Hello'  // <-- params defined in the params-file overrides the ones in the config file
-        result.params.beta == 'World'   // <--   as above
-        result.params.gamma == 'Hello'  // <--   as above
-        result.params.omega == 'Last'
-        result.params.delta == 'Foo'
-        result.params.theta == "$baseDir/something"
-        result.process.publishDir == [path: 'Hello']
-
-        cleanup:
-        file?.delete()
-        paramsFile?.delete()
-    }
-
-    def 'params should override params-file and override params in the config file' () {
-        setup:
-        def params = Files.createTempFile('test', '.yml')
-        params.text = '''
-            alpha: "Hello"
-            beta: "World"
-            omega: "Last"
-            '''.stripIndent()
-        and:
-        def file = Files.createTempFile('test',null)
-        file.text = '''
-        params {
-          alpha = 'x'
-        }
-        params.beta = 'y'
-        params.delta = 'Foo'
-        params.gamma = "I'm gamma"
-        params.omega = "I'm the last"
-
-        process {
-          publishDir = [path: params.alpha]
-        }
-        '''
-        when:
-        def result = configWithParams(file, [paramsFile: params, params: [alpha: 'Hola', beta: 'Mundo']])
-
-        then:
-        result.params.alpha == 'Hola'   // <-- this comes from the CLI
-        result.params.beta == 'Mundo'   // <-- this comes from the CLI as well
-        result.params.omega == 'Last'   // <-- this comes from the params-file
-        result.params.gamma == "I'm gamma"   // <-- from the config
-        result.params.delta == 'Foo'         // <-- from the config
-        result.process.publishDir == [path: 'Hola']
-
-        cleanup:
-        file?.delete()
-        params?.delete()
-    }
-
-    def 'valid config files' () {
-
-        given:
-        def folder = Files.createTempDirectory('test')
-        def f1 = folder.resolve('file1')
-        def f2 = folder.resolve('file2')
-
-        when:
-        new ConfigBuilder()
-                    .validateConfigFiles([f1.toString(), f2.toString()])
-        then:
-        thrown(AbortOperationException)
-
-
-        when:
-        f1.text = '1'; f2.text = '2'
-        def files = new ConfigBuilder()
-                    .validateConfigFiles([f1.toString(), f2.toString()])
-        then:
-        files == [f1, f2]
-
-
-        when:
-        files = new ConfigBuilder(homeDir: folder, currentDir: folder)
-                    .setUserConfigFiles(f1,f2)
-                    .validateConfigFiles()
-        then:
-        files == [f1, f2]
-
-        cleanup:
-        folder.deleteDir()
-
-    }
-
-    def 'should discover default config files' () {
-        given:
-        def homeDir = Files.createTempDirectory('home')
-        def baseDir = Files.createTempDirectory('work')
-        def workDir = Files.createTempDirectory('work')
-
-        when:
-        def homeConfig = homeDir.resolve('config')
-        homeConfig.text = 'foo=1'
-        def files1 = new ConfigBuilder(homeDir: homeDir, baseDir: workDir, currentDir: workDir).validateConfigFiles()
-        then:
-        files1 == [homeConfig]
-
-        when:
-        def workConfig = workDir.resolve('nextflow.config')
-        workConfig.text = 'bar=2'
-        def files2 = new ConfigBuilder(homeDir: homeDir, baseDir: workDir, currentDir: workDir).validateConfigFiles()
-        then:
-        files2 == [homeConfig, workConfig]
-
-        when:
-        def baseConfig = baseDir.resolve('nextflow.config')
-        baseConfig.text = 'ciao=3'
-        def files3 = new ConfigBuilder(homeDir: homeDir, baseDir: baseDir, currentDir: workDir).validateConfigFiles()
-        then:
-        files3 == [homeConfig, baseConfig, workConfig]
-
-
-        cleanup:
-        homeDir?.deleteDir()
-        workDir?.deleteDir()
-    }
-
-    def 'command executor options'() {
-
-        when:
-        def opt = new CliOptions()
-        def run = new CmdRun(executorOptions: [ alpha: 1, 'beta.x': 'hola', 'beta.y': 'ciao' ])
-        def result = new ConfigBuilder().setOptions(opt).setCmdRun(run).buildGivenFiles()
-        then:
-        result.executor.alpha == 1
-        result.executor.beta.x == 'hola'
-        result.executor.beta.y == 'ciao'
-
-    }
-
-    def 'run command cluster options'() {
-
-        when:
-        def opt = new CliOptions()
-        def run = new CmdRun(clusterOptions: [ alpha: 1, 'beta.x': 'hola', 'beta.y': 'ciao' ])
-        def result = new ConfigBuilder().setOptions(opt).setCmdRun(run).buildGivenFiles()
-        then:
-        result.cluster.alpha == 1
-        result.cluster.beta.x == 'hola'
-        result.cluster.beta.y == 'ciao'
-
-    }
-
-    def 'run with docker'() {
-
-        when:
-        def opt = new CliOptions()
-        def run = new CmdRun(withDocker: 'cbcrg/piper')
-        def config = new ConfigBuilder().setOptions(opt).setCmdRun(run).build()
-
-        then:
-        config.docker.enabled
-        config.process.container == 'cbcrg/piper'
-
-    }
-
-    def 'run with docker 2'() {
-
-        given:
-        def file = Files.createTempFile('test','config')
-        file.deleteOnExit()
-        file.text =
-                '''
-                docker {
-                    image = 'busybox'
-                    enabled = false
-                }
-                '''
-
-        when:
-        def opt = new CliOptions(config: [file.toFile().canonicalPath] )
-        def run = new CmdRun(withDocker: '-')
-        def config = new ConfigBuilder().setOptions(opt).setCmdRun(run).build()
-        then:
-        config.docker.enabled
-        config.docker.image == 'busybox'
-        config.process.container == 'busybox'
-
-        when:
-        opt = new CliOptions(config: [file.toFile().canonicalPath] )
-        run = new CmdRun(withDocker: 'cbcrg/mybox')
-        config = new ConfigBuilder().setOptions(opt).setCmdRun(run).build()
-        then:
-        config.docker.enabled
-        config.process.container == 'cbcrg/mybox'
-
-    }
-
-    def 'run with docker 3'() {
-        given:
-        def file = Files.createTempFile('test','config')
-        file.deleteOnExit()
-
-        when:
-        file.text =
-                '''
-                process.'withName:test'.container = 'busybox'
-                '''
-        def opt = new CliOptions(config: [file.toFile().canonicalPath])
-        def run = new CmdRun(withDocker: '-')
-        def config = new ConfigBuilder().setOptions(opt).setCmdRun(run).build()
-        then:
-        config.docker.enabled
-        config.process.'withName:test'.container == 'busybox'
-
-        when:
-        file.text =
-                '''
-                process.container = 'busybox'
-                '''
-        opt = new CliOptions(config: [file.toFile().canonicalPath])
-        run = new CmdRun(withDocker: '-')
-        config = new ConfigBuilder().setOptions(opt).setCmdRun(run).build()
-        then:
-        config.docker.enabled
-        config.process.container == 'busybox'
-
-        when:
-        opt = new CliOptions()
-        run = new CmdRun(withDocker: '-')
-        config = new ConfigBuilder().setOptions(opt).setCmdRun(run).build()
-        then:
-        config.docker.enabled
-        !config.process.container
-
-        when:
-        file.text =
-                '''
-                process.'withName:test'.tag = 'tag'
-                '''
-        opt = new CliOptions(config: [file.toFile().canonicalPath])
-        run = new CmdRun(withDocker: '-')
-        config = new ConfigBuilder().setOptions(opt).setCmdRun(run).build()
-        then:
-        config.docker.enabled
-        !config.process.container
-
-    }
-
-    def 'run without docker'() {
-
-        given:
-        def file = Files.createTempFile('test','config')
-        file.deleteOnExit()
-        file.text =
-                '''
-                docker {
-                    image = 'busybox'
-                    enabled = true
-                }
-                '''
-
-        when:
-        def opt = new CliOptions(config: [file.toFile().canonicalPath] )
-        def run = new CmdRun(withoutDocker: true)
-        def config = new ConfigBuilder().setOptions(opt).setCmdRun(run).build()
-        then:
-        !config.docker.enabled
-        config.docker.image == 'busybox'
-        !config.process.container
-
-    }
-
-    def 'config with cluster options'() {
-
-        when:
-        def opt = new CliOptions()
-        def cmd = new CmdNode(clusterOptions: [join: 'x', group: 'y', interface: 'abc', slots: 10, 'tcp.alpha':'uno', 'tcp.beta': 'due'])
-
-        def config = new ConfigBuilder()
-                .setOptions(opt)
-                .setCmdNode(cmd)
-                .build()
-
-        then:
-        config.cluster.join == 'x'
-        config.cluster.group == 'y'
-        config.cluster.interface == 'abc'
-        config.cluster.slots == 10
-        config.cluster.tcp.alpha == 'uno'
-        config.cluster.tcp.beta == 'due'
-
-    }
-
-    def 'should set session trace options' () {
-
-        given:
-        def env = [:]
-        def builder = [:] as ConfigBuilder
-
-        when:
-        def config = new ConfigObject()
-        builder.configRunOptions(config, env, new CmdRun())
-
-        then:
-        config.trace instanceof Map
-        !config.trace.enabled
-        !config.trace.file
-
-        when:
-        config = new ConfigObject()
-        builder.configRunOptions(config, env, new CmdRun(withTrace: 'some-file'))
-        then:
-        config.trace instanceof Map
-        config.trace.enabled
-        config.trace.file == 'some-file'
-
-        when:
-        config = new ConfigObject()
-        config.trace.file = 'foo.txt'
-        builder.configRunOptions(config, env, new CmdRun(withTrace: 'bar.txt'))
-        then: // command line should override the config file
-        config.trace instanceof Map
-        config.trace.enabled
-        config.trace.file == 'bar.txt'
-
-        when:
-        config = new ConfigObject()
-        config.trace.file = 'foo.txt'
-        builder.configRunOptions(config, env, new CmdRun(withTrace: '-'))
-        then: // command line should override the config file
-        config.trace instanceof Map
-        config.trace.enabled
-        config.trace.file == 'foo.txt'
-
-        when:
-        config = new ConfigObject()
-        builder.configRunOptions(config, env, new CmdRun(withTrace: '-'))
-        then: // command line should override the config file
-        config.trace instanceof Map
-        config.trace.enabled
-        !config.trace.file
-    }
-
-    def 'should set session report options' () {
-
-        given:
-        def env = [:]
-        def builder = [:] as ConfigBuilder
-
-        when:
-        def config = new ConfigObject()
-        builder.configRunOptions(config, env, new CmdRun())
-        then:
-        !config.report
-
-        when:
-        config = new ConfigObject()
-        config.report.file = 'foo.html'
-        builder.configRunOptions(config, env, new CmdRun())
-        then:
-        config.report instanceof Map
-        !config.report.enabled
-        config.report.file == 'foo.html'
-
-        when:
-        config = new ConfigObject()
-        builder.configRunOptions(config, env, new CmdRun(withReport: 'my-report.html'))
-        then:
-        config.report instanceof Map
-        config.report.enabled
-        config.report.file == 'my-report.html'
-
-        when:
-        config = new ConfigObject()
-        config.report.file = 'this-report.html'
-        builder.configRunOptions(config, env, new CmdRun(withReport: 'my-report.html'))
-        then:
-        config.report instanceof Map
-        config.report.enabled
-        config.report.file == 'my-report.html'
-
-        when:
-        config = new ConfigObject()
-        config.report.file = 'this-report.html'
-        builder.configRunOptions(config, env, new CmdRun(withReport: '-'))
-        then:
-        config.report instanceof Map
-        config.report.enabled
-        config.report.file == 'this-report.html'
-
-        when:
-        config = new ConfigObject()
-        builder.configRunOptions(config, env, new CmdRun(withReport: '-'))
-        then:
-        config.report instanceof Map
-        config.report.enabled
-        !config.report.file
-    }
-
-
-    def 'should set session dag options' () {
-
-        given:
-        def env = [:]
-        def builder = [:] as ConfigBuilder
-
-        when:
-        def config = new ConfigObject()
-        builder.configRunOptions(config, env, new CmdRun())
-        then:
-        !config.dag
-
-        when:
-        config = new ConfigObject()
-        config.dag.file = 'foo-dag.html'
-        builder.configRunOptions(config, env, new CmdRun())
-        then:
-        config.dag instanceof Map
-        !config.dag.enabled
-        config.dag.file == 'foo-dag.html'
-
-        when:
-        config = new ConfigObject()
-        builder.configRunOptions(config, env, new CmdRun(withDag: 'my-dag.html'))
-        then:
-        config.dag instanceof Map
-        config.dag.enabled
-        config.dag.file == 'my-dag.html'
-
-        when:
-        config = new ConfigObject()
-        config.dag.file = 'this-dag.html'
-        builder.configRunOptions(config, env, new CmdRun(withDag: 'my-dag.html'))
-        then:
-        config.dag instanceof Map
-        config.dag.enabled
-        config.dag.file == 'my-dag.html'
-
-        when:
-        config = new ConfigObject()
-        config.dag.file = 'this-dag.html'
-        builder.configRunOptions(config, env, new CmdRun(withDag: '-'))
-        then:
-        config.dag instanceof Map
-        config.dag.enabled
-        config.dag.file == 'this-dag.html'
-
-        when:
-        config = new ConfigObject()
-        builder.configRunOptions(config, env, new CmdRun(withDag: '-'))
-        then:
-        config.dag instanceof Map
-        config.dag.enabled
-        !config.dag.file
-    }
-
-    def 'should set session weblog options' () {
-
-        given:
-        def env = [:]
-        def builder = [:] as ConfigBuilder
-
-        when:
-        def config = new ConfigObject()
-        builder.configRunOptions(config, env, new CmdRun())
-        then:
-        !config.weblog
-
-        when:
-        config = new ConfigObject()
-        config.weblog.url = 'http://bar.com'
-        builder.configRunOptions(config, env, new CmdRun())
-        then:
-        config.weblog instanceof Map
-        !config.weblog.enabled
-        config.weblog.url == 'http://bar.com'
-
-        when:
-        config = new ConfigObject()
-        builder.configRunOptions(config, env, new CmdRun(withWebLog: 'http://foo.com'))
-        then:
-        config.weblog instanceof Map
-        config.weblog.enabled
-        config.weblog.url == 'http://foo.com'
-
-        when:
-        config = new ConfigObject()
-        config.weblog.enabled = true
-        config.weblog.url = 'http://bar.com'
-        builder.configRunOptions(config, env, new CmdRun(withWebLog: 'http://foo.com'))
-        then:
-        config.weblog instanceof Map
-        config.weblog.enabled
-        config.weblog.url == 'http://foo.com'
-
-        when:
-        config = new ConfigObject()
-        config.weblog.enabled = true
-        config.weblog.url = 'http://bar.com'
-        builder.configRunOptions(config, env, new CmdRun(withWebLog: '-'))
-        then:
-        config.weblog instanceof Map
-        config.weblog.enabled
-        config.weblog.url == 'http://bar.com'
-
-        when:
-        config = new ConfigObject()
-        config.weblog.enabled = true
-        builder.configRunOptions(config, env, new CmdRun(withWebLog: '-'))
-        then:
-        config.weblog instanceof Map
-        config.weblog.enabled
-        config.weblog.url == 'http://localhost'
-
-    }
-
-    def 'should set session timeline options' () {
-
-        given:
-        def env = [:]
-        def builder = [:] as ConfigBuilder
-
-        when:
-        def config = new ConfigObject()
-        builder.configRunOptions(config, env, new CmdRun())
-        then:
-        !config.timeline
-
-        when:
-        config = new ConfigObject()
-        config.timeline.file = 'my-file.html'
-        builder.configRunOptions(config, env, new CmdRun())
-        then:
-        config.timeline instanceof Map
-        !config.timeline.enabled
-        config.timeline.file == 'my-file.html'
-
-        when:
-        config = new ConfigObject()
-        builder.configRunOptions(config, env, new CmdRun(withTimeline: 'my-timeline.html'))
-        then:
-        config.timeline instanceof Map
-        config.timeline.enabled
-        config.timeline.file == 'my-timeline.html'
-
-        when:
-        config = new ConfigObject()
-        config.timeline.enabled = true
-        config.timeline.file = 'this-timeline.html'
-        builder.configRunOptions(config, env, new CmdRun(withTimeline: 'my-timeline.html'))
-        then:
-        config.timeline instanceof Map
-        config.timeline.enabled
-        config.timeline.file == 'my-timeline.html'
-
-        when:
-        config = new ConfigObject()
-        config.timeline.enabled = true
-        config.timeline.file = 'my-timeline.html'
-        builder.configRunOptions(config, env, new CmdRun(withTimeline: '-'))
-        then:
-        config.timeline instanceof Map
-        config.timeline.enabled
-        config.timeline.file == 'my-timeline.html'
-
-        when:
-        config = new ConfigObject()
-        config.timeline.enabled = true
-        builder.configRunOptions(config, env, new CmdRun(withTimeline: '-'))
-        then:
-        config.timeline instanceof Map
-        config.timeline.enabled
-        !config.timeline.file
-    }
-
-    def 'should set tower options' () {
-
-        given:
-        def env = [:]
-        def builder = [:] as ConfigBuilder
-
-        when:
-        def config = new ConfigObject()
-        builder.configRunOptions(config, env, new CmdRun())
-        then:
-        !config.tower
-
-        when:
-        config = new ConfigObject()
-        config.tower.endpoint = 'http://foo.com'
-        builder.configRunOptions(config, env, new CmdRun())
-        then:
-        config.tower instanceof Map
-        !config.tower.enabled
-        config.tower.endpoint == 'http://foo.com'
-
-        when:
-        config = new ConfigObject()
-        builder.configRunOptions(config, env, new CmdRun(withTower: 'http://bar.com'))
-        then:
-        config.tower instanceof Map
-        config.tower.enabled
-        config.tower.endpoint == 'http://bar.com'
-
-        when:
-        config = new ConfigObject()
-        config.tower.endpoint = 'http://foo.com'
-        builder.configRunOptions(config, env, new CmdRun(withTower: '-'))
-        then:
-        config.tower instanceof Map
-        config.tower.enabled
-        config.tower.endpoint == 'http://foo.com'
-
-        when:
-        config = new ConfigObject()
-        builder.configRunOptions(config, env, new CmdRun(withTower: '-'))
-        then:
-        config.tower instanceof Map
-        config.tower.enabled
-        !config.tower.endpoint
-    }
-
-    def 'should set wave options' () {
-
-        given:
-        def env = [:]
-        def builder = [:] as ConfigBuilder
-
-        when:
-        def config = new ConfigObject()
-        builder.configRunOptions(config, env, new CmdRun())
-        then:
-        !config.wave
-
-        when:
-        config = new ConfigObject()
-        config.wave.endpoint = 'http://foo.com'
-        builder.configRunOptions(config, env, new CmdRun())
-        then:
-        config.wave instanceof Map
-        !config.wave.enabled
-        config.wave.endpoint == 'http://foo.com'
-
-        when:
-        config = new ConfigObject()
-        builder.configRunOptions(config, env, new CmdRun(withWave: 'http://bar.com'))
-        then:
-        config.wave instanceof Map
-        config.wave.enabled
-        config.wave.endpoint == 'http://bar.com'
-
-        when:
-        config = new ConfigObject()
-        config.wave.endpoint = 'http://foo.com'
-        builder.configRunOptions(config, env, new CmdRun(withWave: '-'))
-        then:
-        config.wave instanceof Map
-        config.wave.enabled
-        config.wave.endpoint == 'http://foo.com'
-
-        when:
-        config = new ConfigObject()
-        builder.configRunOptions(config, env, new CmdRun(withWave: '-'))
-        then:
-        config.wave instanceof Map
-        config.wave.enabled
-        !config.wave.endpoint
-    }
-
-    def 'should set cloudcache options' () {
-
-        given:
-        def env = [:]
-        def builder = [:] as ConfigBuilder
-
-        when:
-        def config = new ConfigObject()
-        builder.configRunOptions(config, env, new CmdRun())
-        then:
-        !config.cloudcache
-
-        when:
-        config = new ConfigObject()
-        config.cloudcache.path = 's3://foo/bar'
-        builder.configRunOptions(config, env, new CmdRun())
-        then:
-        config.cloudcache instanceof Map
-        !config.cloudcache.enabled
-        config.cloudcache.path == 's3://foo/bar'
-
-        when:
-        config = new ConfigObject()
-        builder.configRunOptions(config, env, new CmdRun(cloudCachePath: 's3://this/that'))
-        then:
-        config.cloudcache instanceof Map
-        config.cloudcache.enabled
-        config.cloudcache.path == 's3://this/that'
-
-        when:
-        config = new ConfigObject()
-        builder.configRunOptions(config, env, new CmdRun(cloudCachePath: '-'))
-        then:
-        config.cloudcache instanceof Map
-        config.cloudcache.enabled
-        !config.cloudcache.path
-
-        when:
-        config = new ConfigObject()
-        config.cloudcache.path = 's3://alpha/delta'
-        builder.configRunOptions(config, env, new CmdRun(cloudCachePath: '-'))
-        then:
-        config.cloudcache instanceof Map
-        config.cloudcache.enabled
-        config.cloudcache.path == 's3://alpha/delta'
-
-        when:
-        config = new ConfigObject()
-        config.cloudcache.path = 's3://alpha/delta'
-        builder.configRunOptions(config, env, new CmdRun(cloudCachePath: 's3://should/override/config'))
-        then:
-        config.cloudcache instanceof Map
-        config.cloudcache.enabled
-        config.cloudcache.path == 's3://should/override/config'
-
-        when:
-        config = new ConfigObject()
-        config.cloudcache.enabled = false
-        builder.configRunOptions(config, env, new CmdRun(cloudCachePath: 's3://should/override/config'))
-        then:
-        config.cloudcache instanceof Map
-        !config.cloudcache.enabled
-        config.cloudcache.path == 's3://should/override/config'
-
-        when:
-        config = new ConfigObject()
-        builder.configRunOptions(config, [NXF_CLOUDCACHE_PATH:'s3://foo'], new CmdRun(cloudCachePath: 's3://should/override/env'))
-        then:
-        config.cloudcache instanceof Map
-        config.cloudcache.enabled
-        config.cloudcache.path == 's3://should/override/env'
-
-        when:
-        config = new ConfigObject()
-        config.cloudcache.path = 's3://config/path'
-        builder.configRunOptions(config, [NXF_CLOUDCACHE_PATH:'s3://foo'], new CmdRun())
-        then:
-        config.cloudcache instanceof Map
-        config.cloudcache.enabled
-        config.cloudcache.path == 's3://config/path'
-
-        when:
-        config = new ConfigObject()
-        config.cloudcache.path = 's3://config/path'
-        builder.configRunOptions(config, [NXF_CLOUDCACHE_PATH:'s3://foo'], new CmdRun(cloudCachePath: 's3://should/override/config'))
-        then:
-        config.cloudcache instanceof Map
-        config.cloudcache.enabled
-        config.cloudcache.path == 's3://should/override/config'
-
-    }
-
-    def 'should enable conda env' () {
-
-        given:
-        def env = [:]
-        def builder = [:] as ConfigBuilder
-
-        when:
-        def config = new ConfigObject()
-        builder.configRunOptions(config, env, new CmdRun())
-        then:
-        !config.conda
-
-        when:
-        config = new ConfigObject()
-        config.conda.createOptions = 'something'
-        builder.configRunOptions(config, env, new CmdRun())
-        then:
-        config.conda instanceof Map
-        !config.conda.enabled
-        config.conda.createOptions == 'something'
-
-        when:
-        config = new ConfigObject()
-        builder.configRunOptions(config, env, new CmdRun(withConda: 'my-recipe.yml'))
-        then:
-        config.conda instanceof Map
-        config.conda.enabled
-        config.process.conda == 'my-recipe.yml'
-
-        when:
-        config = new ConfigObject()
-        config.conda.enabled = true
-        builder.configRunOptions(config, env, new CmdRun(withConda: 'my-recipe.yml'))
-        then:
-        config.conda instanceof Map
-        config.conda.enabled
-        config.process.conda == 'my-recipe.yml'
-
-        when:
-        config = new ConfigObject()
-        config.process.conda = 'my-recipe.yml'
-        builder.configRunOptions(config, env, new CmdRun(withConda: '-'))
-        then:
-        config.conda instanceof Map
-        config.conda.enabled
-        config.process.conda == 'my-recipe.yml'
-    }
-
-    def 'should disable conda env' () {
-        given:
-        def file = Files.createTempFile('test','config')
-        file.deleteOnExit()
-        file.text =
-                '''
-                conda {
-                    enabled = true
-                }
-                '''
-
-        when:
-        def opt = new CliOptions(config: [file.toFile().canonicalPath] )
-        def run = new CmdRun(withoutConda: true)
-        def config = new ConfigBuilder().setOptions(opt).setCmdRun(run).build()
-        then:
-        !config.conda.enabled
-        !config.process.conda
-    }
-
-    def 'should enable spack env' () {
-
-        given:
-        def env = [:]
-        def builder = [:] as ConfigBuilder
-
-        when:
-        def config = new ConfigObject()
-        builder.configRunOptions(config, env, new CmdRun())
-        then:
-        !config.spack
-
-        when:
-        config = new ConfigObject()
-        config.spack.createOptions = 'something'
-        builder.configRunOptions(config, env, new CmdRun())
-        then:
-        config.spack instanceof Map
-        !config.spack.enabled
-        config.spack.createOptions == 'something'
-
-        when:
-        config = new ConfigObject()
-        builder.configRunOptions(config, env, new CmdRun(withSpack: 'my-recipe.yaml'))
-        then:
-        config.spack instanceof Map
-        config.spack.enabled
-        config.process.spack == 'my-recipe.yaml'
-
-        when:
-        config = new ConfigObject()
-        config.spack.enabled = true
-        builder.configRunOptions(config, env, new CmdRun(withSpack: 'my-recipe.yaml'))
-        then:
-        config.spack instanceof Map
-        config.spack.enabled
-        config.process.spack == 'my-recipe.yaml'
-
-        when:
-        config = new ConfigObject()
-        config.process.spack = 'my-recipe.yaml'
-        builder.configRunOptions(config, env, new CmdRun(withSpack: '-'))
-        then:
-        config.spack instanceof Map
-        config.spack.enabled
-        config.process.spack == 'my-recipe.yaml'
-    }
-
-    def 'should disable spack env' () {
-        given:
-        def file = Files.createTempFile('test','config')
-        file.deleteOnExit()
-        file.text =
-                '''
-                spack {
-                    enabled = true
-                }
-                '''
-
-        when:
-        def opt = new CliOptions(config: [file.toFile().canonicalPath] )
-        def run = new CmdRun(withoutSpack: true)
-        def config = new ConfigBuilder().setOptions(opt).setCmdRun(run).build()
-        then:
-        !config.spack.enabled
-        !config.process.spack
-    }
-
-    def 'SHOULD SET `RESUME` OPTION'() {
-
-        given:
-        def env = [:]
-        def builder = [:] as ConfigBuilder
-
-        when:
-        def config = new ConfigObject()
-        builder.configRunOptions(config, env, new CmdRun())
-        then:
-        !config.isSet('resume')
-
-        when:
-        config = new ConfigObject()
-        config.resume ='alpha-beta-delta'
-        builder.configRunOptions(config, env, new CmdRun())
-        then:
-        config.resume == 'alpha-beta-delta'
-
-        when:
-        config = new ConfigObject()
-        config.resume ='alpha-beta-delta'
-        builder.configRunOptions(config, env, new CmdRun(resume: 'xxx-yyy'))
-        then:
-        config.resume == 'xxx-yyy'
-
-        when:
-        config = new ConfigObject()
-        config.resume ='this-that'
-        builder.configRunOptions(config, env, new CmdRun(resume: 'xxx-yyy'))
-        then:
-        config.resume == 'xxx-yyy'
-    }
-
-    def 'should set `workDir`' () {
-
-        given:
-        def config = new ConfigObject()
-        def builder = [:] as ConfigBuilder
-
-        when:
-        builder.configRunOptions(config, [:], new CmdRun())
-        then:
-        config.workDir == 'work'
-
-        when:
-        config = new ConfigObject()
-        builder.configRunOptions(config, [NXF_WORK: '/foo/bar'], new CmdRun())
-        then:
-        config.workDir == '/foo/bar'
-
-        when:
-        config = new ConfigObject()
-        config.workDir = 'hello/there'
-        builder.configRunOptions(config, [:], new CmdRun())
-        then:
-        config.workDir == 'hello/there'
-
-        when:
-        config = new ConfigObject()
-        config.workDir = 'hello/there'
-        builder.configRunOptions(config, [:], new CmdRun(workDir: 'my/work/dir'))
-        then:
-        config.workDir == 'my/work/dir'
-    }
-
-    def 'should set `libDir`' () {
-        given:
-        def config = new ConfigObject()
-        def builder = [:] as ConfigBuilder
-
-        when:
-        builder.configRunOptions(config, [:], new CmdRun())
-        then:
-        !config.isSet('libDir')
-
-        when:
-        builder.configRunOptions(config, [NXF_LIB:'/foo/bar'], new CmdRun())
-        then:
-        config.libDir == '/foo/bar'
-
-        when:
-        builder.configRunOptions(config, [:], new CmdRun(libPath: 'my/lib/dir'))
-        then:
-        config.libDir == 'my/lib/dir'
-    }
-
-    def 'should set `cacheable`' () {
-        given:
-        def env = [:]
-        def config
-        def builder = [:] as ConfigBuilder
-
-        when:
-        config = new ConfigObject()
-        builder.configRunOptions(config, env, new CmdRun())
-        then:
-        !config.isSet('cacheable')
-
-        when:
-        config = new ConfigObject()
-        builder.configRunOptions(config, env, new CmdRun(cacheable: false))
-        then:
-        config.cacheable == false
-
-        when:
-        config = new ConfigObject()
-        builder.configRunOptions(config, env, new CmdRun(cacheable: true))
-        then:
-        config.cacheable == true
     }
 
     def 'should check for a valid profile' () {
@@ -1533,7 +211,6 @@ class ConfigBuilderTest extends Specification {
             """
             .stripIndent().leftTrim()
 
-
         when:
         builder.profile = 'delta'
         builder.checkValidProfile(['alpha','delta','omega'])
@@ -1555,156 +232,118 @@ class ConfigBuilderTest extends Specification {
 
     }
 
-    def 'should set profile options' () {
-
-        def builder
-
-        when:
-        builder = new ConfigBuilder().setCmdRun(new CmdRun(profile: 'foo'))
-        then:
-        builder.profile == 'foo'
-        builder.validateProfile
-
-        when:
-        builder = new ConfigBuilder().setCmdRun(new CmdRun())
-        then:
-        builder.profile == 'standard'
-        !builder.validateProfile
-
-        when:
-        builder = new ConfigBuilder().setCmdRun(new CmdRun(profile: 'standard'))
-        then:
-        builder.profile == 'standard'
-        builder.validateProfile
-    }
-
-    def 'should set config options' () {
-        def builder
-
-        when:
-        builder = new ConfigBuilder().setCmdConfig(new CmdConfig())
-        then:
-        !builder.showAllProfiles
-
-        when:
-        builder = new ConfigBuilder().setCmdConfig(new CmdConfig(showAllProfiles: true))
-        then:
-        builder.showAllProfiles
-
-        when:
-        builder = new ConfigBuilder().setCmdConfig(new CmdConfig(profile: 'foo'))
-        then:
-        builder.profile == 'foo'
-        builder.validateProfile
-
-    }
-
-    def 'should set params into config object' () {
+    def 'should warn about missing attribute' () {
 
         given:
-        def emptyFile = Files.createTempFile('empty','config').toFile()
-        def EMPTY = [emptyFile.toString()]
-
-        def configFile = Files.createTempFile('test','config').toFile()
-        configFile.deleteOnExit()
-        configFile.text = '''
-          params.foo = 1
-          params.bar = 2
-          params.data = '/some/path'
-        '''
-        configFile = configFile.toString()
-
-        def jsonFile = Files.createTempFile('test','.json').toFile()
-        jsonFile.text = '''
-          {
-            "foo": 10,
-            "bar": 20
-          }
-        '''
-        jsonFile = jsonFile.toString()
-
-        def yamlFile = Files.createTempFile('test','.yaml').toFile()
-        yamlFile.text = '''
-          {
-            "foo": 100,
-            "bar": 200
-          }
-        '''
-        yamlFile = yamlFile.toString()
-
-        def config
+        SysEnv.push('NXF_SYNTAX_PARSER': 'v1', HOME: '/home/user')
+        def file = Files.createTempFile('test','config')
+        file.deleteOnExit()
+        file.text =
+                '''
+                params.foo = HOME
+                '''
 
         when:
-        config = configWithParams([:], [:], [config: EMPTY])
+        def cfg = new ConfigBuilder().build([file])
         then:
-        config.params == [:]
+        cfg.params.foo == '/home/user'
 
-        // get params for the CLI
         when:
-        config = configWithParams([:], [params: [foo:'one', bar:'two']], [config: EMPTY])
+        file.text =
+                '''
+                params.foo = bar
+                '''
+        new ConfigBuilder().build([file])
         then:
-        config.params == [foo:'one', bar:'two']
+        def e = thrown(ConfigParseException)
+        e.message == "Unknown config attribute `bar` -- check config file: ${file.toRealPath()}".toString()
 
-        // get params from config file
-        when:
-        config = configWithParams([:], [:], [config: [configFile]])
-        then:
-        config.params == [foo:1, bar:2, data: '/some/path']
-
-        // get params form JSON file
-        when:
-        config = configWithParams([:], [paramsFile: jsonFile], [config: EMPTY])
-        then:
-        config.params == [foo:10, bar:20]
-
-        // get params from YAML file
-        when:
-        config = configWithParams([:], [paramsFile: yamlFile], [config: EMPTY])
-        then:
-        config.params == [foo:100, bar:200]
-
-        // cli override config
-        when:
-        config = configWithParams([:], [params: [foo:'hello', baz:'world']], [config: [configFile]])
-        then:
-        config.params == [foo:'hello', bar:2, baz: 'world', data: '/some/path']
-
-        // CLI override JSON
-        when:
-        config = configWithParams([:], [params: [foo:'hello', baz:'world'], paramsFile: jsonFile], [config: EMPTY])
-        then:
-        config.params == [foo:'hello', bar:20, baz: 'world']
-
-        // JSON override config
-        when:
-        config = configWithParams([:], [paramsFile: jsonFile], [config: [configFile]])
-        then:
-        config.params == [foo:10, bar:20, data: '/some/path']
-
-
-        // CLI override JSON that override config
-        when:
-        config = configWithParams([:], [paramsFile: jsonFile, params: [foo:'Ciao']], [config: [configFile]])
-        then:
-        config.params == [foo:'Ciao', bar:20, data: '/some/path']
+        cleanup:
+        SysEnv.pop()
     }
 
-    def 'should run with conda' () {
-
+    def 'should guarantee top scopes without a session scope' () {
         when:
-        def config = new ConfigBuilder().setCmdRun(new CmdRun(withConda: '/some/path/env.yml')).build()
+        def config = new ConfigBuilder().build([])
         then:
-        config.process.conda == '/some/path/env.yml'
-
+        config.containsKey('env')
+        config.containsKey('process')
+        config.containsKey('executor')
+        config.containsKey('params')
+        and:
+        // `session` is not a config scope and must not be created implicitly
+        !config.containsKey('session')
     }
 
-    def 'should run with spack' () {
+    def 'should not expose environment variables to the config binding with v2 parser' () {
+        given:
+        SysEnv.push('NXF_SYNTAX_PARSER': 'v2', HOME: '/home/user')
+        def file = Files.createTempFile('test','config')
+        file.deleteOnExit()
+        file.text =
+                '''
+                params.foo = HOME
+                '''
 
         when:
-        def config = new ConfigBuilder().setCmdRun(new CmdRun(withSpack: '/some/path/env.yaml')).build()
+        // unlike the v1 parser, the v2 parser does not bind environment variables,
+        // so `HOME` is an unknown attribute rather than resolving to the env value
+        new ConfigBuilder().build([file])
         then:
-        config.process.spack == '/some/path/env.yaml'
+        thrown(ConfigParseException)
 
+        cleanup:
+        SysEnv.pop()
+    }
+
+    def 'should render missing variables' () {
+        given:
+        SysEnv.push('NXF_SYNTAX_PARSER': 'v1')
+        def file = Files.createTempFile('test',null)
+
+        file.text =
+                '''
+                foo = 'xyz'
+                bar = "$SCRATCH/singularity_images_nextflow"
+                '''
+
+        when:
+        def builder = new ConfigBuilder()
+                .setShowMissingVariables(true)
+        def cfg = builder.build([file])
+        def str = ConfigHelper.toCanonicalString(cfg)
+        then:
+        str == '''\
+            foo = 'xyz'
+            bar = '$SCRATCH/singularity_images_nextflow'
+            '''.stripIndent()
+
+        and:
+        builder.warnings[0].startsWith('Unknown config attribute `SCRATCH`')
+        cleanup:
+        SysEnv.pop()
+        file?.delete()
+    }
+
+    def 'should report fully qualified missing attribute'  () {
+
+        given:
+        SysEnv.push('NXF_SYNTAX_PARSER': 'v1')
+        def file = Files.createTempFile('test','config')
+        file.deleteOnExit()
+
+        when:
+        file.text =
+                '''
+                params.x = foo.bar
+                '''
+        new ConfigBuilder().build([file])
+        then:
+        def e = thrown(ConfigParseException)
+        e.message == "Unknown config attribute `foo.bar` -- check config file: ${file.toRealPath()}".toString()
+
+        cleanup:
+        SysEnv.pop()
     }
 
     def 'should collect config files' () {
@@ -1732,91 +371,6 @@ class ConfigBuilderTest extends Specification {
         file2?.delete()
     }
 
-
-    def 'should configure notification' () {
-
-        given:
-        Map config
-
-        when:
-        config = new ConfigBuilder().setCmdRun(new CmdRun()).build()
-        then:
-        !config.notification
-
-        when:
-        config = new ConfigBuilder().setCmdRun(new CmdRun(withNotification: true)).build()
-        then:
-        config.notification.enabled == true
-
-        when:
-        config = new ConfigBuilder().setCmdRun(new CmdRun(withNotification: false)).build()
-        then:
-        config.notification.enabled == false
-        config.notification.to == null
-
-        when:
-        config = new ConfigBuilder().setCmdRun(new CmdRun(withNotification: 'yo@nextflow.com')).build()
-        then:
-        config.notification.enabled == true
-        config.notification.to == 'yo@nextflow.com'
-    }
-
-    def 'should configure fusion' () {
-
-        given:
-        Map config
-
-        when:
-        config = new ConfigBuilder().setCmdRun(new CmdRun()).build()
-        then:
-        !config.fusion
-
-        when:
-        config = new ConfigBuilder().setCmdRun(new CmdRun(withFusion: true)).build()
-        then:
-        config.fusion.enabled == true
-
-        when:
-        config = new ConfigBuilder().setCmdRun(new CmdRun(withFusion: false)).build()
-        then:
-        config.fusion == [enabled: false]
-
-        when:
-        config = new ConfigBuilder().setCmdRun(new CmdRun(withFusion: true)).build()
-        then:
-        config.fusion == [enabled: true]
-    }
-
-    def 'should configure stub run mode' () {
-        given:
-        Map config
-
-        when:
-        config = new ConfigBuilder().setCmdRun(new CmdRun()).build()
-        then:
-        !config.stubRun
-
-        when:
-        config = new ConfigBuilder().setCmdRun(new CmdRun(stubRun: true)).build()
-        then:
-        config.stubRun == true
-    }
-
-    def 'should configure preview mode' () {
-        given:
-        Map config
-
-        when:
-        config = new ConfigBuilder().setCmdRun(new CmdRun()).build()
-        then:
-        !config.preview
-
-        when:
-        config = new ConfigBuilder().setCmdRun(new CmdRun(preview: true)).build()
-        then:
-        config.preview == true
-    }
-
     def 'should merge profiles' () {
         given:
         def ENV = [:]
@@ -1826,7 +380,7 @@ class ConfigBuilderTest extends Specification {
         def CONFIG = '''
                 process.container = 'base'
                 process.executor = 'local'
-
+                
                 profiles {
                     cfg1 {
                       process.executor = 'sge'
@@ -1853,7 +407,7 @@ class ConfigBuilderTest extends Specification {
 
         when:
         builder.profile = 'cfg1'
-        result = builder.buildConfig0(ENV, [CONFIG])
+        result = builder.build(ENV, [CONFIG])
         then:
         result.process.container == 'base'
         result.process.executor == 'sge'
@@ -1861,7 +415,7 @@ class ConfigBuilderTest extends Specification {
 
         when:
         builder.profile = 'cfg2'
-        result = builder.buildConfig0(ENV, [CONFIG])
+        result = builder.build(ENV, [CONFIG])
         then:
         result.process.container == 'base'
         result.process.executor == 'batch'
@@ -1869,7 +423,7 @@ class ConfigBuilderTest extends Specification {
 
         when:
         builder.profile = 'cfg1,docker'
-        result = builder.buildConfig0(ENV, [CONFIG])
+        result = builder.build(ENV, [CONFIG])
         then:
         result.process.container ==  'foo/1'
         result.process.executor == 'sge'
@@ -1879,7 +433,7 @@ class ConfigBuilderTest extends Specification {
 
         when:
         builder.profile = 'cfg1,singularity'
-        result = builder.buildConfig0(ENV, [CONFIG])
+        result = builder.build(ENV, [CONFIG])
         then:
         result.process.container ==  'bar-2.img'
         result.process.executor == 'sge'
@@ -1889,7 +443,7 @@ class ConfigBuilderTest extends Specification {
 
         when:
         builder.profile = 'cfg2,singularity'
-        result = builder.buildConfig0(ENV, [CONFIG])
+        result = builder.build(ENV, [CONFIG])
         then:
         result.process.container ==  'bar-2.img'
         result.process.executor == 'batch'
@@ -1899,7 +453,7 @@ class ConfigBuilderTest extends Specification {
 
         when:
         builder.profile = 'missing'
-        builder.buildConfig0(ENV, [CONFIG])
+        builder.build(ENV, [CONFIG])
         then:
         thrown(AbortOperationException)
     }
@@ -1938,7 +492,7 @@ class ConfigBuilderTest extends Specification {
 
         when:
         builder.showAllProfiles = true
-        result = builder.buildConfig0(ENV, [CONFIG])
+        result = builder.build(ENV, [CONFIG])
         then:
         result.profiles.cfg1.process.executor == 'sge'
         result.profiles.cfg1.process.queue == 'short'
@@ -1955,7 +509,7 @@ class ConfigBuilderTest extends Specification {
 
         given:
         def folder = Files.createTempDirectory('test')
-        def file1 = folder.resolve('test.config')
+        def file1 = folder.resolve('test.conf')
         file1.text = '''
             process {
                 cpus = 2
@@ -1971,7 +525,7 @@ class ConfigBuilderTest extends Specification {
             '''
 
         when:
-        def cfg1 = new ConfigBuilder().buildConfig0([:], [file1])
+        def cfg1 = new ConfigBuilder().build([file1])
         then:
         cfg1.process.cpus == 2
         cfg1.process.'withName:bar'.cpus == 4
@@ -1982,10 +536,10 @@ class ConfigBuilderTest extends Specification {
         when:
         def file2 = folder.resolve('nextflow.config')
         file2.text = """
-            includeConfig "$file1"
+            includeConfig "$file1" 
             """
 
-        def cfg2 = new ConfigBuilder().buildConfig0([:], [file2])
+        def cfg2 = new ConfigBuilder().build([file2])
         then:
         cfg1 == cfg2
 
@@ -1998,20 +552,20 @@ class ConfigBuilderTest extends Specification {
 
         given:
         def folder = Files.createTempDirectory('test')
-        def file1 = folder.resolve('test.config')
+        def file1 = folder.resolve('test.conf')
         file1.text = '''
             process {
-                ext { args = "Hello World!" }
-                cpus = 1
+                ext { args = "Hello World!" } 
+                cpus = 1 
                 withName:BAR {
-                    ext { args = "Ciao mondo!" }
+                    ext { args = "Ciao mondo!" } 
                     cpus = 2
                 }
             }
             '''
 
         when:
-        def cfg1 = new ConfigBuilder().buildConfig0([:], [file1])
+        def cfg1 = new ConfigBuilder().build([file1])
         then:
         cfg1.process.cpus == 1
         cfg1.process.ext.args == 'Hello World!'
@@ -2029,11 +583,11 @@ class ConfigBuilderTest extends Specification {
 
         given:
         def folder = Files.createTempDirectory('test')
-        def file1 = folder.resolve('test.config')
+        def file1 = folder.resolve('test.conf')
         file1.text = '''
             process {
-                ext.args = "Hello World!"
-                cpus = 1
+                ext.args = "Hello World!" 
+                cpus = 1 
                 withName:BAR {
                     ext.args = "Ciao mondo!"
                     cpus = 2
@@ -2042,7 +596,7 @@ class ConfigBuilderTest extends Specification {
             '''
 
         when:
-        def cfg1 = new ConfigBuilder().buildConfig0([:], [file1])
+        def cfg1 = new ConfigBuilder().build([file1])
         then:
         cfg1.process.cpus == 1
         cfg1.process.ext.args == 'Hello World!'
@@ -2056,18 +610,18 @@ class ConfigBuilderTest extends Specification {
     def 'should access top params from profile' () {
         given:
         def folder = Files.createTempDirectory('test')
-        def file1 = folder.resolve('file1.config')
+        def file1 = folder.resolve('file1.conf')
 
         file1.text = """
-            params.alpha = 1
-            params.delta = 2
-
+            params.alpha = 1 
+            params.delta = 2 
+            
             profiles {
                 foo {
-                    params.delta = 20
-                    params.gamma = 30
-
-                    process {
+                    params.delta = 20 
+                    params.gamma = 30 
+                    
+                    process {  
                         cpus = params.alpha
                     }
                 }
@@ -2075,7 +629,7 @@ class ConfigBuilderTest extends Specification {
             """
 
         when:
-        def cfg = new ConfigBuilder().setProfile('foo').buildConfig0([:], [file1])
+        def cfg = new ConfigBuilder().setProfile('foo').build([file1])
         then:
         cfg.process == [cpus: 1]
         cfg.params == [alpha: 1, delta: 20, gamma: 30]
@@ -2087,31 +641,30 @@ class ConfigBuilderTest extends Specification {
     def 'should access top params from profile [2]' () {
         given:
         def folder = Files.createTempDirectory('test')
-        def file1 = folder.resolve('file1.config')
+        def file1 = folder.resolve('file1.conf')
 
         file1.text = """
             params {
-                alpha = 1
-                delta = 2
+                alpha = 1 
+                delta = 2 
             }
-
+            
             profiles {
                 foo {
                     params {
-                        delta = 20
+                        delta = 20 
                         gamma = 30
                     }
 
-                    process {
+                    process {  
                         cpus = params.alpha
                     }
                 }
             }
             """
 
-
         when:
-        def cfg = new ConfigBuilder().setProfile('foo').buildConfig0([:], [file1])
+        def cfg = new ConfigBuilder().setProfile('foo').build([file1])
         then:
         cfg.process == [cpus: 1]
         cfg.params == [alpha: 1, delta: 20, gamma: 30]
@@ -2123,33 +676,32 @@ class ConfigBuilderTest extends Specification {
     def 'should merge params two profiles' () {
         given:
         def folder = Files.createTempDirectory('test')
-        def file1 = folder.resolve('file1.config')
+        def file1 = folder.resolve('file1.conf')
 
-        file1.text = '''
+        file1.text = '''    
             profiles {
                 foo {
                     params {
-                        alpha = 1
-                        delta = 2
+                        alpha = 1 
+                        delta = 2 
                     }
                 }
 
                 bar {
                     params {
-                        delta = 20
-                        gamma = 30
+                        delta = 20 
+                        gamma = 30 
                     }
-
+                    
                     process {
                         cpus = params.alpha
-                    }
+                    }                
                 }
             }
             '''
 
-
         when:
-        def cfg = new ConfigBuilder().setProfile('foo,bar') .buildConfig0([:], [file1])
+        def cfg = new ConfigBuilder().setProfile('foo,bar') .build([file1])
         then:
         cfg.process.cpus == 1
         cfg.params.alpha == 1
@@ -2158,236 +710,6 @@ class ConfigBuilderTest extends Specification {
 
         cleanup:
         folder?.deleteDir()
-    }
-
-    def 'CLI params should overwrite only the key provided when nested'() {
-        given:
-        def folder = File.createTempDir()
-        def configMain = new File(folder, 'nextflow.config').absoluteFile
-
-        configMain.text = """
-        params {
-            foo = 'Hello'
-            bar = "Monde"
-            baz {
-                x = "Ciao"
-                y = "mundo"
-                z {
-                    alpha = "Hallo"
-                    beta  = "World"
-                }
-            }
-
-        }
-        """
-
-        when:
-        def config = configWithParams(
-            [env: [NXF_CONFIG_FILE: configMain.toString()]],
-            [params: [bar: "world", 'baz.y': "mondo", 'baz.z.beta': "Welt"]] )
-
-        then:
-        config.params.foo == 'Hello'
-        config.params.bar == 'world'
-        config.params.baz.x == 'Ciao'
-        config.params.baz.y == 'mondo'
-        //tests recursion
-        config.params.baz.z.alpha == 'Hallo'
-        config.params.baz.z.beta == 'Welt'
-
-        cleanup:
-        folder?.deleteDir()
-    }
-
-    @Unroll
-    def 'should merge config params' () {
-        given:
-        def builder = new ConfigBuilder()
-
-        expect:
-        def cfg = new ConfigObject(); if(CONFIG) cfg.putAll(CONFIG)
-        and:
-        builder.mergeMaps(cfg, PARAMS, false) == EXPECTED
-
-        where:
-        CONFIG                      | PARAMS | EXPECTED
-        [foo:1]                     | null                          | [foo:1]
-        null                        | [bar:2]                       | [bar:2]
-        [foo:1]                     | [bar:2]                       | [foo: 1, bar: 2]
-        [foo:1]                     | [bar:null]                    | [foo: 1, bar: null]
-        [foo:1]                     | [foo:null]                    | [foo: null]
-        [foo:1, bar:[:]]            | [bar: [x:10, y:20]]           | [foo: 1, bar: [x:10, y:20]]
-        [foo:1, bar:[x:1, y:2]]     | [bar: [x:10, y:20]]           | [foo: 1, bar: [x:10, y:20]]
-        [foo:1, bar:[x:1, y:2]]     | [foo: 2, bar: [x:10, y:20]]   | [foo: 2, bar: [x:10, y:20]]
-        [foo:1, bar:null]           | [bar: [x:10, y:20]]           | [foo: 1, bar: [x:10, y:20]]
-        [foo:1, bar:2]              | [bar: [x:10, y:20]]           | [foo: 1, bar: [x:10, y:20]]
-        [foo:1, bar:[x:1, y:2]]     | [bar: 2]                      | [foo: 1, bar: 2]
-    }
-
-    @Unroll
-    def 'should merge config strict params' () {
-        given:
-        def builder = new ConfigBuilder()
-
-        expect:
-        def cfg = new ConfigObject(); if(CONFIG) cfg.putAll(CONFIG)
-        and:
-        builder.mergeMaps(cfg, PARAMS, true) == EXPECTED
-
-        where:
-        CONFIG                      | PARAMS                        | EXPECTED
-        [:]                         | [bar:2]                       | [bar:2]
-        [foo:1]                     | null                          | [foo:1]
-        null                        | [bar:2]                       | [bar:2]
-        [foo:1]                     | [bar:2]                       | [foo: 1, bar: 2]
-        [foo:1]                     | [bar:null]                    | [foo: 1, bar: null]
-        [foo:1]                     | [foo:null]                    | [foo: null]
-        [foo:1, bar:[:]]            | [bar: [x:10, y:20]]           | [foo: 1, bar: [x:10, y:20]]
-        [foo:1, bar:[x:1, y:2]]     | [bar: [x:10, y:20]]           | [foo: 1, bar: [x:10, y:20]]
-        [foo:1, bar:[x:1, y:2]]     | [foo: 2, bar: [x:10, y:20]]   | [foo: 2, bar: [x:10, y:20]]
-    }
-
-    def 'prevent side effects with with nested params' () {
-        given:
-        def folder = Files.createTempDirectory('test')
-        and:
-        def config = folder.resolve('nf.config')
-        config.text = '''\
-        params.test.foo = "foo_def"
-        params.test.bar = "bar_def"
-        '''.stripIndent()
-
-        when:
-        def cfg1 = new ConfigBuilder()
-                .setOptions( new CliOptions(userConfig: [config.toString()]))
-                .build()
-        then:
-        cfg1.params.test.foo == "foo_def"
-        cfg1.params.test.bar == "bar_def"
-
-
-        when:
-        def cfg2 = configWithParams([:], [params: ['test.foo': 'CLI_FOO']], [userConfig: [config.toString()]])
-        then:
-        cfg2.params.test.foo == "CLI_FOO"
-        cfg2.params.test.bar == "bar_def"
-
-        cleanup:
-        folder?.deleteDir()
-    }
-
-    def 'parse nested json' () {
-        given:
-        def folder = Files.createTempDirectory('test')
-        and:
-        def config = folder.resolve('nf.json')
-        config.text = '''\
-        {
-            "title": "something",
-            "nested": {
-                "name": "Mike",
-                "and": {
-                    "more": "nesting",
-                    "still": {
-                        "another": "layer"
-                    }
-                }
-            }
-        }
-        '''.stripIndent()
-
-        when:
-        def cfg1 = configWithParams([:], [paramsFile: config.toString()])
-
-        then:
-        cfg1.params.title == "something"
-        cfg1.params.nested.name == 'Mike'
-        cfg1.params.nested.and.more == 'nesting'
-        cfg1.params.nested.and.still.another == 'layer'
-
-        cleanup:
-        folder?.deleteDir()
-    }
-
-    def 'parse nested yaml' () {
-        given:
-        def folder = Files.createTempDirectory('test')
-        and:
-        def config = folder.resolve('nf.yaml')
-        config.text = '''\
-            title: "something"
-            nested:
-              name: "Mike"
-              and:
-                more: nesting
-                still:
-                  another: layer
-        '''.stripIndent()
-
-        when:
-        def cfg1 = configWithParams([:], [paramsFile: config.toString()])
-
-        then:
-        cfg1.params.title == "something"
-        cfg1.params.nested.name == 'Mike'
-        cfg1.params.nested.and.more == 'nesting'
-        cfg1.params.nested.and.still.another == 'layer'
-
-        cleanup:
-        folder?.deleteDir()
-    }
-
-
-    def 'should return parsed config' () {
-        given:
-        def base = Files.createTempDirectory('test')
-        def configFile = base.resolve('nextflow.config')
-        configFile.text = '''
-        profiles {
-            first {
-                params {
-                  foo = 'Hello world'
-                  awsKey = 'xyz'
-                }
-                process {
-                    executor = { 'local' }
-                }
-                outputDir = params.outdir
-            }
-            second {
-                params.none = 'Blah'
-            }
-        }
-        '''
-        when:
-        def opt = new CliOptions(config: [configFile.toFile().canonicalPath])
-        def cmd = new CmdRun(profile: 'first', withTower: 'http://foo.com', launcher: new Launcher(options: opt))
-        def cliParams = [foo: 'Hola', outdir: 'output_folder']
-        def txt = ConfigBuilder.resolveConfig(base, cmd, cliParams)
-        then:
-        txt == '''\
-            params {
-               foo = 'Hola'
-               outdir = 'output_folder'
-               awsKey = '[secret]'
-            }
-
-            process {
-               executor = { 'local' }
-            }
-
-            outputDir = 'output_folder'
-            outputFormat = 'text'
-            workDir = 'work'
-
-            tower {
-               enabled = true
-               endpoint = 'http://foo.com'
-            }
-            '''.stripIndent()
-
-        cleanup:
-        base?.deleteDir()
     }
 
     def 'should merge profiles with conditions' () {
@@ -2428,7 +750,7 @@ class ConfigBuilderTest extends Specification {
         '''
 
         when:
-        def cfg = new ConfigBuilder().setProfile('test').buildConfig0([:], [main])
+        def cfg = new ConfigBuilder().setProfile('test').build([:], [main])
         then:
         cfg.process.'withName:FOO'
         cfg.params.load_config == true
@@ -2438,7 +760,6 @@ class ConfigBuilderTest extends Specification {
         cleanup:
         folder?.deleteDir()
     }
-
 
     def 'should build config object with secrets' () {
         given:
@@ -2465,7 +786,7 @@ class ConfigBuilderTest extends Specification {
         '''
 
         when:
-        def cfg = new ConfigBuilder().setBaseDir(folder).buildConfig0([:], [text])
+        def cfg = new ConfigBuilder().setBaseDir(folder).build([text])
         then:
         cfg.params.p == "$folder/1"
         cfg.params.s == 'ciao/2'
@@ -2514,7 +835,7 @@ class ConfigBuilderTest extends Specification {
 
         snippet1.text = """
         p2 = secrets.DELTA
-        includeConfig "$snippet2"
+        includeConfig "$snippet2" 
         """
 
         snippet2.text = '''
@@ -2523,8 +844,7 @@ class ConfigBuilderTest extends Specification {
         '''
 
         when:
-        def opt = new CliOptions()
-        def config = new ConfigBuilder().setBaseDir(folder).setOptions(opt).buildGivenFiles(configMain)
+        def config = new ConfigBuilder().setBaseDir(folder).build([configMain])
         then:
         config.p1 == 'one'
         config.p2 == 'two'
@@ -2571,7 +891,7 @@ class ConfigBuilderTest extends Specification {
 
         snippet1.text = '''
         p2 = 'two'
-        // include config via string interpolation using a secret
+        // include config via string interpolation using a secret 
         includeConfig "$secrets.SECRET_FILE2"
         '''
 
@@ -2580,8 +900,7 @@ class ConfigBuilderTest extends Specification {
         '''
 
         when:
-        def opt = new CliOptions()
-        def config = new ConfigBuilder().setBaseDir(folder).setOptions(opt).buildGivenFiles(configMain)
+        def config = new ConfigBuilder().setBaseDir(folder).build([configMain])
         then:
         config.p1 == 'one'
         config.p2 == 'two'
@@ -2633,9 +952,9 @@ class ConfigBuilderTest extends Specification {
         and:
         configMain.text = '''
         p1 = secrets.ALPHA
-        foo.p1 = secrets.ALPHA
+        foo.p1 = secrets.ALPHA 
         bar {
-          p1 = secrets.ALPHA
+          p1 = secrets.ALPHA 
         }
         // include config via a secret property
         includeConfig secrets.SECRET_FILE1
@@ -2647,7 +966,7 @@ class ConfigBuilderTest extends Specification {
         bar {
           p2 = "$secrets.DELTA"
         }
-        // include config via string interpolation using a secret
+        // include config via string interpolation using a secret 
         includeConfig "$secrets.SECRET_FILE2"
         '''
 
@@ -2660,12 +979,10 @@ class ConfigBuilderTest extends Specification {
         '''
 
         when:
-        def opt = new CliOptions()
         def config = new ConfigBuilder()
-                            .setBaseDir(folder)
-                            .setOptions(opt)
-                            .setStripSecrets(true)
-                            .buildGivenFiles(configMain)
+                .setBaseDir(folder)
+                .setStripSecrets(true)
+                .build([configMain])
         then:
         config.p1 == 'secrets.ALPHA'
         config.p2 == 'secrets.DELTA'
@@ -2685,4 +1002,3 @@ class ConfigBuilderTest extends Specification {
     }
 
 }
-
