+++ modules/nextflow/src/test/groovy/nextflow/config/parser/v1/ConfigParserV1Test.groovy
@@ -575,7 +575,7 @@ class ConfigParserV1Test extends Specification {
 
         when:
         def url = 'http://localhost:9900/nextflow.config' as Path
-        def cfg = new ConfigBuilder().buildGivenFiles(url)
+        def cfg = new ConfigBuilder().build([url])
         then:
         cfg.params.foo == 'Hello'
         cfg.params.bar == 'world!'
