+++ plugins/nf-tower/src/main/io/seqera/tower/plugin/BaseCommandImpl.groovy
@@ -51,8 +51,9 @@ class BaseCommandImpl {
     }
 
     protected Map readConfig() {
-        final builder = new ConfigBuilder().setHomeDir(Const.APP_HOME_DIR).setCurrentDir(Const.APP_HOME_DIR)
-        return builder.buildConfigObject().flatten()
+        // note: the Nextflow home config file is `$HOME/.nextflow/config` (not `nextflow.config`)
+        final configFile = Const.APP_HOME_DIR.resolve('config')
+        return new ConfigBuilder().build(configFile.exists() ? [ configFile ] : []).flatten()
     }
 
     protected List<Map> listUserWorkspaces(TowerClient client, String userId) {
